+++ src/components/Notion.tsx
@@ -1,24 +1,20 @@
-import NotionCustomBlock from '../interfaces/NotionCustomBlock'
 import NotionQueryData from '../interfaces/NotionQueryData'
+import NotionRenderConfig from '../interfaces/config/NotionRenderConfig'
 import NotionBlock from './NotionBlock'
 
 interface Props {
   data: NotionQueryData
-  custom?: NotionCustomBlock[]
+  config?: NotionRenderConfig
 }
 
-export default function Notion({ data, custom }: Props) {
-  return (
-    <>
-      {data.map((object) => (
-        <NotionBlock
-          block={object.block}
-          children={object.children}
-          level={object.level}
-          key={object.block.id}
-          custom={custom}
-        />
-      ))}
-    </>
-  )
+export default function Notion({ data, config }: Readonly<Props>) {
+  return data.map((object) => (
+    <NotionBlock
+      block={object.block}
+      sub={object.sub}
+      level={object.level}
+      key={object.block.id}
+      config={config}
+    />
+  ))
 }