+++ src/helper.tsx
@@ -1,7 +1,0 @@
-import NotionRichText from './components/common/NotionRichText'
-import getNotionImageURL from './methods/data/getImageURL'
-
-import type NotionQueryData from './interfaces/NotionQueryData'
-
-export { NotionRichText, getNotionImageURL }
-export type { NotionQueryData }