+++ src/index.ts
@@ -1,0 +1,12 @@
+export { default as NotionRichText } from './components/common/NotionRichText'
+export { default as NotionBlock } from './components/NotionBlock'
+export { default as NotionQuery } from './classes/NotionQuery'
+export { default as NotionParser } from './classes/NotionParser'
+
+export type { default as NotionQueryData } from './interfaces/NotionQueryData'
+export type { default as NotionBlockObject } from './interfaces/NotionBlockObject'
+export type { default as NotionBlockTypes } from './interfaces/NotionBlockTypes'
+export type { default as NotionCustomBlock } from './interfaces/NotionCustomBlock'
+export type { default as NotionRenderConfig } from './interfaces/config/NotionRenderConfig'
+
+export { default } from './components/Notion'