+++ src/interfaces/NotionCustomBlock.ts
@@ -3,6 +3,13 @@ import NotionBlockObject from './NotionBlockObject'
 import NotionBlockTypes from './NotionBlockTypes'
 
 export default interface NotionCustomBlock {
+  /**
+   * The type of the Notion block.
+   */
   type: NotionBlockTypes
+
+  /**
+   * The React component used to render the Notion block.
+   */
   component: FunctionComponent<NotionBlockObject<any>>
 }