+++ src/interfaces/NotionQueryData.ts
@@ -2,4 +2,7 @@ import NotionBlockObject from './NotionBlockObject'
 
 type NotionQueryData = NotionBlockObject[]
 
+/**
+ * Represents the data structure for a Notion query result, which is an array of Notion block objects.
+ */
 export default NotionQueryData