+++ src/interfaces/NotionQueryOptions.ts
@@ -1,3 +1,6 @@
 export default interface NotionQueryOptions {
+  /**
+   * Enable debug mode to log additional information during the Notion query process.
+   */
   debug?: boolean
 }